refactored fastparse parser and tests
