+++ parser/src/main/scala/eo/parser/fastparse/Tokens.scala
@@ -1,0 +1,46 @@
+package eo.parser.fastparse
+
+import eo.core.ast.{EOCharData, EOExpr, EOIntData, EOStrData}
+import eo.core.ast.astparams.EOExprOnly
+import fastparse._, NoWhitespace._
+import higherkindness.droste.data.Fix
+
+object Tokens {
+
+  def comment[_: P]: P[Unit] = P(
+    whitespaceChars ~ "#" ~ CharsWhile(_ != '\n', 0) ~ "\n"
+  )
+
+  def identifier[_: P]: P[String] =
+    P(lowercase ~ (letter | digit | "-" | "_").rep).!
+
+  private def letter[_: P] = P(lowercase | uppercase)
+
+  private def lowercase[_: P] = P(CharIn("a-z"))
+
+  private def uppercase[_: P] = P(CharIn("A-Z"))
+
+  def digit[_: P]: P[String] = P(CharIn("0-9")).!
+
+  def integer[_: P]: P[EOExprOnly] = P("-".? ~ CharsWhileIn("0-9")).!
+    .map(int => Fix[EOExpr](EOIntData(int.toInt)))
+
+  def char[_: P]: P[EOExprOnly] = P("\'" ~ letter.! ~ "\'")
+    .map(c => Fix[EOExpr](EOCharData(c.charAt(0))))
+
+  def string[_: P]: P[EOExprOnly] = P("\"" ~ CharsWhile(_ != '\"').! ~ "\"")
+    .map(s => Fix[EOExpr](EOStrData(s)))
+
+  private def whitespaceChars[_: P] = P(
+    CharsWhile(c => c == '\t' || c == ' ' || c == '\r', 0)
+  )
+
+  def emptyLine[_: P]: P[Unit] = P(whitespaceChars ~ "\n")
+
+  def emptyLinesOrComments[_: P]: P[Unit] = P(
+    (emptyLine | comment).rep
+  )
+
+  def singleLineWhitespace[_: P]: P[Unit] =
+    CharsWhileIn(" \t", 1)
+}
