+++ parser/src/test/scala/eo/parser/MutualRecExample.scala
@@ -1,0 +1,137 @@
+package eo.parser
+
+import com.github.tarao.nonempty.collection.NonEmpty
+import eo.core.ast.astparams.EOExprOnly
+import eo.core.ast._
+import higherkindness.droste.data.Fix
+
+object MutualRecExample {
+  val ast: EOProg[EOExprOnly] = EOProg(
+    EOMetas(
+      pack = Some("sandbox"),
+      metas = Vector(
+        EOAliasMeta("stdout", "org.eolang.io.stdout"),
+        EOAliasMeta("sprintf", "org.eolang.txt.sprintf"),
+      )
+    ),
+    Vector(
+      EOBndExpr(
+        EOAnyNameBnd(LazyName("base")),
+        Fix[EOExpr](
+          EOObj(
+            freeAttrs = Vector(),
+            varargAttr = None,
+            bndAttrs = Vector(
+              EOBndExpr(
+                EOAnyNameBnd(LazyName("x")),
+                Fix[EOExpr](EOSimpleApp("memory"))
+              ),
+              EOBndExpr(
+                EOAnyNameBnd(LazyName("f")),
+                Fix[EOExpr](
+                  EOObj(
+                    freeAttrs = Vector(LazyName("self"), LazyName("v")),
+                    varargAttr = None,
+                    bndAttrs = Vector(
+                      EOBndExpr(
+                        EODecoration,
+                        Fix[EOExpr](
+                          EOCopy(
+                            Fix[EOExpr](EODot(Fix[EOExpr](EOSimpleApp("x")), "write")),
+                            NonEmpty[Vector[EOBnd[EOExprOnly]]](
+                              EOAnonExpr(Fix[EOExpr](EOSimpleApp("v")))
+                            )
+                          )
+                        )
+                      )
+                    )
+                  )
+                )
+              ),
+              EOBndExpr(
+                EOAnyNameBnd(LazyName("g")),
+                Fix[EOExpr](
+                  EOObj(
+                    freeAttrs = Vector(LazyName("self"), LazyName("v")),
+                    varargAttr = None,
+                    bndAttrs = Vector(
+                      EOBndExpr(
+                        EODecoration,
+                        Fix[EOExpr](
+                          EOCopy(
+                            Fix[EOExpr](EODot(Fix[EOExpr](EOSimpleApp("self")), "f")),
+                            NonEmpty[Vector[EOBnd[EOExprOnly]]](
+                              EOAnonExpr(Fix[EOExpr](EOSimpleApp("self"))),
+                              EOAnonExpr(Fix[EOExpr](EOSimpleApp("v")))
+                            )
+                          )
+                        )
+                      )
+                    )
+                  )
+                )
+              )
+            )
+          )
+        )
+      ),
+
+
+      EOBndExpr(
+        EOAnyNameBnd(LazyName("derived")),
+        Fix[EOExpr](
+          EOObj(
+            freeAttrs = Vector(),
+            varargAttr = None,
+            bndAttrs = Vector(
+              EOBndExpr(EODecoration, Fix[EOExpr](EOSimpleApp("base"))),
+              EOBndExpr(
+                EOAnyNameBnd(LazyName("f")),
+                Fix[EOExpr](
+                  EOObj(
+                    freeAttrs = Vector(LazyName("self"), LazyName("v")),
+                    varargAttr = None,
+                    bndAttrs = Vector(
+                      EOBndExpr(
+                        EODecoration,
+                        Fix[EOExpr](
+                          EOCopy(
+                            Fix[EOExpr](EODot(Fix[EOExpr](EOSimpleApp("self")), "g")),
+                            NonEmpty[Vector[EOBnd[EOExprOnly]]](
+                              EOAnonExpr(Fix[EOExpr](EOSimpleApp("self"))),
+                              EOAnonExpr(Fix[EOExpr](EOSimpleApp("v")))
+                            )
+                          )
+                        )
+                      )
+                    ),
+                  )
+                )
+              )
+            )
+          )
+        )
+      ),
+    )
+  )
+  val code: String =
+    """+package sandbox
+      |+alias stdout org.eolang.io.stdout
+      |+alias sprintf org.eolang.txt.sprintf
+      |[] > base
+      |  memory > x
+      |  [self v] > f
+      |    x.write > @
+      |      v
+      |  [self v] > g
+      |    self.f > @
+      |      self
+      |      v
+      |[] > derived
+      |  base > @
+      |  [self v] > f
+      |    self.g > @
+      |      self
+      |      v
+      |""".stripMargin
+}
